write down what i could remeber about movie
i used two files in one folder to create "hover" so when the mouse goes over it will highlight. link the two i had to use <link rel="stylesheet" herif="spoiler.css"> i didn't quite get it but i asked someone for help on the link part, so now i got a better grasp of it.